refactor: oauth repository
