+++ server/src/types.ts
@@ -8,6 +8,7 @@ import { ConfigRepository } from 'src/repositories/config.repository';
 import { LoggingRepository } from 'src/repositories/logging.repository';
 import { MediaRepository } from 'src/repositories/media.repository';
 import { MemoryRepository } from 'src/repositories/memory.repository';
+import { OAuthRepository } from 'src/repositories/oauth.repository';
 import { VersionHistoryRepository } from 'src/repositories/version-history.repository';
 import { ViewRepository } from 'src/repositories/view-repository';
 
@@ -42,6 +43,7 @@ export type ILoggingRepository = Pick<
 >;
 export type IMediaRepository = RepositoryInterface<MediaRepository>;
 export type IMemoryRepository = RepositoryInterface<MemoryRepository>;
+export type IOAuthRepository = RepositoryInterface<OAuthRepository>;
 export type IViewRepository = RepositoryInterface<ViewRepository>;
 export type IVersionHistoryRepository = RepositoryInterface<VersionHistoryRepository>;
